DevOps_basics_202104/Exam/vagrant/docker/app/producer/code/requirements.txt
mysql-connector-python==8.0.25
protobuf==3.17.3